Move 'InMemoryFeedStore' to its own file in production
